# cc1/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type DmsUserRole = typeof DmsUserRole[keyof typeof DmsUserRole];


export const DmsUserRole = {
  admin: 'admin',
  staff: 'staff',
} as const;

export interface DmsUser {
  id: number;
  name: string;
  email: string;
  role: DmsUserRole;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: DmsUser;
}

export type DmsUserInputRole = typeof DmsUserInputRole[keyof typeof DmsUserInputRole];


export const DmsUserInputRole = {
  admin: 'admin',
  staff: 'staff',
} as const;

export interface DmsUserInput {
  name: string;
  email: string;
  password: string;
  role: DmsUserInputRole;
}

export type DmsUserUpdateRole = typeof DmsUserUpdateRole[keyof typeof DmsUserUpdateRole];


export const DmsUserUpdateRole = {
  admin: 'admin',
  staff: 'staff',
} as const;

export interface DmsUserUpdate {
  name?: string;
  email?: string;
  role?: DmsUserUpdateRole;
}

export interface Project {
  id: number;
  name: string;
  type: string;
  description: string;
  createdAt: string;
  updatedAt?: string;
}

export interface ProjectWithStats {
  id: number;
  name: string;
  type: string;
  description: string;
  createdAt: string;
  updatedAt?: string;
  documentCount: number;
}

export interface Document {
  id: number;
  title: string;
  fileName: string;
  fileType: string;
  fileSize: number;
  projectId: number;
  /** @nullable */
  projectName?: string | null;
  uploadedBy: number;
  uploadedByName: string;
  createdAt: string;
}

export interface ProjectWithDocuments {
  id: number;
  name: string;
  type: string;
  description: string;
  createdAt: string;
  updatedAt?: string;
  documents: Document[];
}

export interface ProjectInput {
  name: string;
  type: string;
  description: string;
}

export interface ProjectUpdate {
  name?: string;
  type?: string;
  description?: string;
}

export interface DocumentUploadInput {
  title: string;
  projectId: number;
}

export interface DownloadInfo {
  url: string;
  fileName: string;
}

export interface ProjectTypeStat {
  type: string;
  count: number;
}

export interface DashboardStats {
  totalProjects: number;
  totalDocuments: number;
  totalUsers: number;
  recentUploads: Document[];
  projectsByType: ProjectTypeStat[];
}

export interface SearchResults {
  projects: ProjectWithStats[];
  documents: Document[];
}

export type DmsListProjectsParams = {
/**
 * Search by project name
 */
q?: string;
};

export type DmsListDocumentsParams = {
projectId?: number;
q?: string;
};

export type DmsSearchParams = {
q: string;
};